# pricing/shoppers-pricing-openapi-v1.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Shoppers Catalogue server.\n\n[Learn about Swagger](http://swagger.io) or join the IRC channel `#swagger` on irc.freenode.net.\n\nFor this sample, you can use the api key `special-key` to test the authorization filters\n",
        "version": "1.0.0",
        "title": "Shoppers OpenAPI",
        "termsOfService": "http://helloreverb.com/terms/",
        "contact": {
            "name": "harjeetss@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "shoppers.io",
    "basePath": "/v2",
    "schemes": [
        "http"
    ],
    "paths": {
        "/v1/products/{sku}/group-prices/{customerGroupId}/tiers": {
            "get": {
                "tags": [
                    "catalogProductTierPriceManagementV1"
                ],
                "description": "Get tier price of product",
                "operationId": "catalogProductTierPriceManagementV1GetListGet",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "customerGroupId",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "description": "'all' can be used to specify 'ALL GROUPS'"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "200 Success.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/product-tier-price-interface"
                            }
                        }
                    },
                    "400": {
                        "description": "400 Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    }
                }
            }
        },
        "/v1/products/{sku}/group-prices/{customerGroupId}/tiers/{qty}/price/{price}": {
            "post": {
                "tags": [
                    "catalogProductTierPriceManagementV1"
                ],
                "description": "Create tier price for product",
                "operationId": "catalogProductTierPriceManagementV1AddPost",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "customerGroupId",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "description": "'all' can be used to specify 'ALL GROUPS'"
                    },
                    {
                        "name": "price",
                        "in": "path",
                        "type": "number",
                        "required": true
                    },
                    {
                        "name": "qty",
                        "in": "path",
                        "type": "number",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "200 Success.",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "400 Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    }
                }
            }
        },
        "/v1/products/{sku}/group-prices/{customerGroupId}/tiers/{qty}": {
            "delete": {
                "tags": [
                    "catalogProductTierPriceManagementV1"
                ],
                "description": "Remove tier price from product",
                "operationId": "catalogProductTierPriceManagementV1RemoveDelete",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "customerGroupId",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "description": "'all' can be used to specify 'ALL GROUPS'"
                    },
                    {
                        "name": "qty",
                        "in": "path",
                        "type": "number",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "200 Success.",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "400 Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error-response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "error-response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "description": "Error message"
                },
                "errors": {
                    "$ref": "#/definitions/error-errors"
                },
                "code": {
                    "type": "integer",
                    "description": "Error code"
                },
                "parameters": {
                    "$ref": "#/definitions/error-parameters"
                },
                "trace": {
                    "type": "string",
                    "description": "Stack trace"
                }
            },
            "required": [
                "message"
            ]
        },
        "error-errors": {
            "type": "array",
            "description": "Errors list",
            "items": {
                "$ref": "#/definitions/error-errors-item"
            }
        },
        "error-errors-item": {
            "type": "object",
            "description": "Error details",
            "properties": {
                "message": {
                    "type": "string",
                    "description": "Error message"
                },
                "parameters": {
                    "$ref": "#/definitions/error-parameters"
                }
            }
        },
        "error-parameters": {
            "type": "array",
            "description": "Error parameters list",
            "items": {
                "$ref": "#/definitions/error-parameters-item"
            }
        },
        "error-parameters-item": {
            "type": "object",
            "description": "Error parameters item",
            "properties": {
                "resources": {
                    "type": "string",
                    "description": "ACL resource"
                },
                "fieldName": {
                    "type": "string",
                    "description": "Missing or invalid field name"
                },
                "fieldValue": {
                    "type": "string",
                    "description": "Incorrect field value"
                }
            }
        },
        "framework-attribute-interface": {
            "type": "object",
            "description": "Interface for custom attribute value.",
            "properties": {
                "attribute_code": {
                    "type": "string",
                    "description": "Attribute code"
                },
                "value": {
                    "type": "string",
                    "description": "Attribute value"
                }
            },
            "required": [
                "attribute_code",
                "value"
            ]
        },
        "product": {
            "type": "object",
            "description": "",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "Id"
                },
                "sku": {
                    "type": "string",
                    "description": "Sku"
                },
                "name": {
                    "type": "string",
                    "description": "Name"
                },
                "attribute_set_id": {
                    "type": "integer",
                    "description": "Attribute set id"
                },
                "price": {
                    "type": "number",
                    "description": "Price"
                },
                "status": {
                    "type": "integer",
                    "description": "Status"
                },
                "visibility": {
                    "type": "integer",
                    "description": "Visibility"
                },
                "type_id": {
                    "type": "string",
                    "description": "Type id"
                },
                "created_at": {
                    "type": "string",
                    "description": "Created date"
                },
                "updated_at": {
                    "type": "string",
                    "description": "Updated date"
                },
                "weight": {
                    "type": "number",
                    "description": "Weight"
                },
                "extension_attributes": {
                    "$ref": "#/definitions/product-extension-interface"
                },
                "product_links": {
                    "type": "array",
                    "description": "Product links info",
                    "items": {
                        "$ref": "#/definitions/product-link-interface"
                    }
                },
                "options": {
                    "type": "array",
                    "description": "List of product options",
                    "items": {
                        "$ref": "#/definitions/product-custom-option-interface"
                    }
                },
                "tier_prices": {
                    "type": "array",
                    "description": "List of product tier prices",
                    "items": {
                        "$ref": "#/definitions/product-tier-price-interface"
                    }
                },
                "custom_attributes": {
                    "type": "array",
                    "description": "Custom attributes values.",
                    "items": {
                        "$ref": "#/definitions/framework-attribute-interface"
                    }
                }
            },
            "required": [
                "sku"
            ]
        },
        "Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "product-extension-interface": {
            "type": "object",
            "description": "extending",
            "properties": {
                "stock_item": {
                    "$ref": "#/definitions/inventory-data-stock-item-interface"
                },
                "bundle_product_options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bundle-data-option-interface"
                    }
                },
                "configurable_product_options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/configurable-product-data-option-interface"
                    }
                },
                "configurable_product_links": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "inventory-data-stock-item-interface": {
            "type": "object",
            "description": "Interface StockItem",
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "stock_id": {
                    "type": "integer",
                    "description": "Stock identifier"
                },
                "qty": {
                    "type": "number"
                },
                "is_in_stock": {
                    "type": "boolean",
                    "description": "Stock Availability"
                },
                "is_qty_decimal": {
                    "type": "boolean"
                },
                "show_default_notification_message": {
                    "type": "boolean"
                },
                "use_config_min_qty": {
                    "type": "boolean"
                },
                "min_qty": {
                    "type": "number",
                    "description": "Minimal quantity available for item status in stock"
                },
                "use_config_min_sale_qty": {
                    "type": "integer"
                },
                "min_sale_qty": {
                    "type": "number",
                    "description": "Minimum Qty Allowed in Shopping Cart or NULL when there is no limitation"
                },
                "use_config_max_sale_qty": {
                    "type": "boolean"
                },
                "max_sale_qty": {
                    "type": "number",
                    "description": "Maximum Qty Allowed in Shopping Cart data wrapper"
                },
                "use_config_backorders": {
                    "type": "boolean"
                },
                "backorders": {
                    "type": "integer",
                    "description": "Backorders status"
                },
                "use_config_notify_stock_qty": {
                    "type": "boolean"
                },
                "notify_stock_qty": {
                    "type": "number",
                    "description": "Notify for Quantity Below data wrapper"
                },
                "use_config_qty_increments": {
                    "type": "boolean"
                },
                "qty_increments": {
                    "type": "number",
                    "description": "Quantity Increments data wrapper"
                },
                "use_config_enable_qty_inc": {
                    "type": "boolean"
                },
                "enable_qty_increments": {
                    "type": "boolean",
                    "description": "Whether Quantity Increments is enabled"
                },
                "use_config_manage_stock": {
                    "type": "boolean"
                },
                "manage_stock": {
                    "type": "boolean",
                    "description": "Can Manage Stock"
                },
                "low_stock_date": {
                    "type": "string"
                },
                "is_decimal_divided": {
                    "type": "boolean"
                },
                "stock_status_changed_auto": {
                    "type": "integer"
                }
            },
            "required": [
                "qty",
                "is_in_stock",
                "is_qty_decimal",
                "show_default_notification_message",
                "use_config_min_qty",
                "min_qty",
                "use_config_min_sale_qty",
                "min_sale_qty",
                "use_config_max_sale_qty",
                "max_sale_qty",
                "use_config_backorders",
                "backorders",
                "use_config_notify_stock_qty",
                "notify_stock_qty",
                "use_config_qty_increments",
                "qty_increments",
                "use_config_enable_qty_inc",
                "enable_qty_increments",
                "use_config_manage_stock",
                "manage_stock",
                "low_stock_date",
                "is_decimal_divided",
                "stock_status_changed_auto"
            ]
        },
        "bundle-data-option-interface": {
            "type": "object",
            "description": "Interface OptionInterface",
            "properties": {
                "option_id": {
                    "type": "integer",
                    "description": "Option id"
                },
                "title": {
                    "type": "string",
                    "description": "Option title"
                },
                "required": {
                    "type": "boolean",
                    "description": "Is required option"
                },
                "type": {
                    "type": "string",
                    "description": "Input type"
                },
                "position": {
                    "type": "integer",
                    "description": "Option position"
                },
                "sku": {
                    "type": "string",
                    "description": "Product sku"
                },
                "product_links": {
                    "type": "array",
                    "description": "Product links",
                    "items": {
                        "$ref": "#/definitions/bundle-data-link-interface"
                    }
                }
            }
        },
        "bundle-data-link-interface": {
            "type": "object",
            "description": "Interface LinkInterface",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "The identifier"
                },
                "sku": {
                    "type": "string",
                    "description": "Linked product sku"
                },
                "option_id": {
                    "type": "integer",
                    "description": "Option id"
                },
                "qty": {
                    "type": "number",
                    "description": "Qty"
                },
                "position": {
                    "type": "integer",
                    "description": "Position"
                },
                "is_default": {
                    "type": "boolean",
                    "description": "Is default"
                },
                "price": {
                    "type": "number",
                    "description": "Price"
                },
                "price_type": {
                    "type": "integer",
                    "description": "Price type"
                },
                "can_change_quantity": {
                    "type": "integer",
                    "description": "Whether quantity could be changed"
                }
            },
            "required": [
                "is_default",
                "price",
                "price_type"
            ]
        },
        "configurable-product-data-option-interface": {
            "type": "object",
            "description": "Interface OptionInterface",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "attribute_id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "is_use_default": {
                    "type": "boolean"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/configurable-product-data-option-value-interface"
                    }
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "configurable-product-data-option-value-interface": {
            "type": "object",
            "description": "Interface OptionValueInterface",
            "properties": {
                "value_index": {
                    "type": "integer"
                }
            },
            "required": [
                "value_index"
            ]
        },
        "product-link-interface": {
            "type": "object",
            "description": "",
            "properties": {
                "sku": {
                    "type": "string",
                    "description": "SKU"
                },
                "link_type": {
                    "type": "string",
                    "description": "Link type"
                },
                "linked_product_sku": {
                    "type": "string",
                    "description": "Linked product sku"
                },
                "linked_product_type": {
                    "type": "string",
                    "description": "Linked product type (simple, virtual, etc)"
                },
                "position": {
                    "type": "integer",
                    "description": "Linked item position"
                }
            },
            "required": [
                "sku",
                "link_type",
                "linked_product_sku",
                "linked_product_type",
                "position"
            ]
        },
        "product-custom-option-interface": {
            "type": "object",
            "description": "",
            "properties": {
                "product_sku": {
                    "type": "string",
                    "description": "Product SKU"
                },
                "option_id": {
                    "type": "integer",
                    "description": "Option id"
                },
                "title": {
                    "type": "string",
                    "description": "Option title"
                },
                "type": {
                    "type": "string",
                    "description": "Option type"
                },
                "sort_order": {
                    "type": "integer",
                    "description": "Sort order"
                },
                "is_require": {
                    "type": "boolean",
                    "description": "Is require"
                },
                "price": {
                    "type": "number",
                    "description": "Price"
                },
                "price_type": {
                    "type": "string",
                    "description": "Price type"
                },
                "sku": {
                    "type": "string",
                    "description": "Sku"
                },
                "file_extension": {
                    "type": "string"
                },
                "max_characters": {
                    "type": "integer"
                },
                "image_size_x": {
                    "type": "integer"
                },
                "image_size_y": {
                    "type": "integer"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product-custom-option-values-interface"
                    }
                }
            },
            "required": [
                "product_sku",
                "title",
                "type",
                "sort_order",
                "is_require"
            ]
        },
        "product-custom-option-values-interface": {
            "type": "object",
            "description": "",
            "properties": {
                "title": {
                    "type": "string",
                    "description": "Option title"
                },
                "sort_order": {
                    "type": "integer",
                    "description": "Sort order"
                },
                "price": {
                    "type": "number",
                    "description": "Price"
                },
                "price_type": {
                    "type": "string",
                    "description": "Price type"
                },
                "sku": {
                    "type": "string",
                    "description": "Sku"
                },
                "option_type_id": {
                    "type": "integer",
                    "description": "Option type id"
                }
            },
            "required": [
                "title",
                "sort_order",
                "price",
                "price_type"
            ]
        },
        "product-tier-price-interface": {
            "type": "object",
            "description": "",
            "properties": {
                "customer_group_id": {
                    "type": "integer",
                    "description": "Customer group id"
                },
                "qty": {
                    "type": "number",
                    "description": "Tier qty"
                },
                "value": {
                    "type": "number",
                    "description": "Price value"
                }
            },
            "required": [
                "customer_group_id",
                "qty",
                "value"
            ]
        }
    }
}